requirements.txt
asyncio-nats-client
msgpack-python
